REfactored into null_output.hpp
